feat(upstream, check): patches the sync applies, and the inference fix one was waiting for
`tools/upstream/sync.sh` checked `upstream/flow` out and applied nothing to it,
so a fix to Meta's Flow Rust port could not be kept anywhere. #205 found the
root cause of its own defect, wrote the diff, measured it, and reverted it for
that reason. #300 and the profiling half of #678 are behind the same wall.

`tools/upstream/patches/flow` is now applied after the checkout, so a checkout
is the pinned commit *plus* those patches — on a laptop, in every CI job, and
after `uf run setup`. The failure it is designed against is not a patch that
does not apply, it is one that is quietly not applied: nothing fails to
compile, no test goes red, and `uf check` answers differently than it did
yesterday.

  * The directory is the list. Every `NNNN-*.patch` is applied in that order,
    and a file in there with any other name is refused rather than ignored,
    because a `0001-x.patch.orig` sitting beside the patches would never be
    applied and would look exactly like one that was.
  * Each patch carries an `Issue:` line above the diff or the sync refuses it.
  * Three questions per patch: the pinned commit already has it (it landed
    upstream — delete the file, and the sync says so), the work tree already
    has it (leave the file *and its mtime* alone, or every sync rebuilds the
    port), or it applies. Anything else stops the sync after restoring the
    submodule, so half-patched is not a state anything can compile.
  * Before that loop, the tree's modified paths are reconciled against the
    paths the patches claim, so a patch deleted from the directory stops being
    carried — the one case a per-patch loop cannot see.

A warm sync with one patch applied costs 0.17 s and writes nothing. The
shallow, blobless, sparse checkout is untouched.

`tools/upstream/test-patches.sh` runs the mechanism against every way a patch
can go missing, on a scratch submodule, and is wired into `ci` — the subject of
this check fails silently, so nothing but the cases says it works. It caught
two real holes while it was being written: a deleted patch that stayed in the
work tree, and a patch that had landed upstream and was kept anyway.

0001 is #205, the diff that had nowhere to live. A `match` pattern binding over
a generic container came out as `unknown` where `switch` on the same tag came
out as the type variable, because `handle_generic` does not list
`ConcretizeForMatchArg` and a `GenericT` therefore fell to the generic-erasing
catch-all in `dispatch.rs` — which sits above the match-argument collector and
below the predicate one. That asymmetry is the whole bug. Letting the
`GenericT` through is necessary and not sufficient: on its own it makes
`match (tag: T)` over `T extends "a" | "b"` report `match-not-exhaustive` with
both arms unused, so the value union also has to remember the variable it
unwrapped and put it back in `to_type`.

    -  Cannot call change with value bound to the first parameter
    -  because unknown [1] is incompatible with empty [2].
    +  no problems

It also removes a false positive from a package this repository ships:
`unwrap<T>` in `@uniflowed/state` matches over `Loadable<T>` and bound `data`
as `unknown`. Nothing else in the type-test corpus moves — 9 fixtures, 143
diagnostics before and 142 after, and that is the one.

`crates/uf_check/tests/known_bugs.rs` keeps #300 and hands #205 to
`crates/uf_check/tests/upstream_patches.rs`, which is where a test that pins a
patch lives: it stays when the patch is deleted, because it is what says the
fix survived the bump. #300 is untouched — a different site, in the conditional
type's tuple matcher, and `tests/type-tests/server-actions.js` still pins it.

Closes #205. Mechanism: #707.
